Add files via upload
This project was developed for the course “Programación 1.” In this application, you can learn university-level Physics I while tracking all your progress statistics. Although the software is designed for a general audience and includes physics-based games to facilitate learning, my partner and I have chosen to offer it as a commercial product. Therefore, you cannot access the source code or executable without purchasing a license. This project is protected by copyright.